requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
#mysqlclient
python-dotenv
PyMySQL